reverse-searcher/requirements.txt
# Core ML and Image Processing
tensorflow==2.15.0
efficientnet==1.1.1
Pillow==10.1.0
numpy==1.24.3

# Vector Search and Storage
faiss-cpu==1.7.4
redis==5.0.1

# Web Framework and API
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-json-logger==2.0.4

# Security and Validation
python-magic==0.4.27
Werkzeug==3.0.1

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0 